Implement conceptual typing throughout the system.
At this point _f is used in many places, but not every place it should be. We will need to review that. Also a few types that aren't conceptual maybe should become concepts, perhaps any class Name_* such that we use a class IsName_f should be implemented with concepts, perhaps not.